final project of nest working

+++ src/modules/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -10,7 +10,8 @@ import * as bcrypt from 'bcrypt';
 export class UserService {
 
   constructor(
-    @InjectRepository(UserEntity) private userRepo: Repository<UserEntity>
+    @InjectRepository(UserEntity) 
+      private userRepo: Repository<UserEntity>
   ) {}
 
   async create(createUserDto: CreateUserDto) {
@@ -29,30 +30,53 @@ export class UserService {
     newUser.email = createUserDto.email;
     newUser.password = password;
     newUser.address = createUserDto.address;
-    // Save the new user in the database
-    return await this.userRepo.save(newUser);
+    // Save the new user in the database and return without password field
+    const ans = await this.userRepo.save(newUser);
+    return { id: ans.id, name: ans.name, surname: ans.surname, email: ans.email, address: ans.address };
+    // await this.userRepo.save(newUser);
+    // return this.userRepo.findOne( {  ,select: ['id', 'name', 'surname', 'email', 'address'] });
   }
 
   // Return all users
   async findAll() {
-    return await this.userRepo.find();
+    // Return all users from the database without password field
+    // return await this.userRepo.find();
+    return await this.userRepo.find({ select: ['id', 'name', 'surname', 'email', 'address'] });
   }
 
   // Find a user by id
   async findOne(id: number) {
-    return await this.userRepo.findOne(id);
+    // Verify if user exists
+    const user = await this.userRepo.findOne({ id: id });
+    if (!user) {
+      // If user doesn't exists, throw an error 404 and return
+      throw new HttpException(`User with id '${id}' not found`, HttpStatus.NOT_FOUND);
+    }
+    // return user;    
+    return { id: user.id, name: user.name, surname: user.surname, email: user.email, address: user.address };
   }
 
   // Update a user
   async update(id: number, updateUserDto: UpdateUserDto) { 
-    const user = await this.userRepo.findOne(id);
+    // Verify if user to update exists
+    const user = await this.userRepo.findOne({ email: updateUserDto.email });      
+    if (!user) {
+      // If user doesn't exists, throw an error 404
+      throw new HttpException('User to update not found', HttpStatus.NOT_FOUND);
+    }
     this.userRepo.merge(user, updateUserDto); 
     return this.userRepo.save(user);
   }
 
-  // Delete a user
+  // Delete a user by id
   async remove(id: number) {
+    // Verify if user to delete exists
+    const user = await this.userRepo.findOne({ id: id });
+    if (!user) {
+      // If user doesn't exists, throw an error 404
+      throw new HttpException('User to delete not found', HttpStatus.NOT_FOUND);
+    }
     await this.userRepo.delete(id);
-    return true;
+    return { id: id, name: user.name, surname: user.surname, email: user.email, address: user.address };
   }
 }
